% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgm_summaries.R
\name{cgm_summaries}
\alias{cgm_summaries}
\title{Computes all metrics}
\usage{
cgm_summaries(
  df_dex,
  start = "default",
  end = "default",
  lookback = 90,
  freq = NULL,
  breaks = NULL,
  include_bv = T
)
}
\arguments{
\item{include_bv}{}
}
\description{
Wrapper function whcih calls the other computation functions to quickly compute all CMG metrics.
}
