% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_agp.R
\name{compute_agp}
\alias{compute_agp}
\title{Compute AGP Thermometers}
\usage{
compute_agp(x, inter, include_bv = T)
}
\arguments{
\item{x}{Data frame with format identical to that outputted by read_dexcom}

\item{inter}{Time interval for computation; e.g. every 14 days, every month, etc.}

\item{include_bv}{"High" and "Low" flags are converted to value 400 and 40
respectively and used in computations}
}
\value{
Data frame with AGP metrics
}
\description{
Compute AGP Thermometers
}
