% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_avg_glucose.R
\name{compute_avg_glucose}
\alias{compute_avg_glucose}
\title{Compute average glucose in given time periods}
\usage{
compute_avg_glucose(df)
}
\arguments{
\item{df}{}
}
\description{
Compute average glucose in given time periods
}
