% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_bv.R
\name{convert_bv}
\alias{convert_bv}
\title{Converts high/low flags to numeric values}
\usage{
convert_bv(x)
}
\arguments{
\item{x}{Data frame, read_dexcom output}
}
\value{
Data frame with bg_value_num updated to include numeric values
associated with high/low flags
}
\description{
Converts high/low flags to numeric values
}
