% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_window.R
\name{partition_window}
\alias{partition_window}
\title{Partition the time window into intervals or breaks}
\usage{
partition_window(df_dexcom, freq = NULL, breaks = NULL)
}
\arguments{
\item{breaks}{}
}
\description{
Partition the time window into intervals or breaks
}
