% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cohort_dexcom.R
\name{read_cohort_dexcom}
\alias{read_cohort_dexcom}
\title{Reads cohort dexcom data}
\usage{
read_cohort_dexcom(folder_path, indiv = F, ordered = T)
}
\arguments{
\item{folder_path}{Path to cohort folder.}

\item{indiv}{Boolean. If (default) FALSE, all patient files are binded into a single data frame.
If TRUE, then each patient data frame is stored in an entry in a list of data frames.
The order of the list corresponds to the order in which patient .csv files were read.}

\item{ordered}{Boolean. Only relevant when 'indiv = FALSE'. If 'ordered = TRUE', then the
single data frame is rearranged into chronological order (as opposed to leaving the grouping of
observations by patient)}
}
\value{
A data frame or a list of data frames.
If 'indiv = F', a data frame in the format of read_dexcom output.
If 'indiv = T', a list of data frame, each in the format of read_dexcom output.
}
\description{
Reads cohort dexcom data
}
\examples{
read_cohort_dexcom("~/Desktop/R Workspace/4T/external files/cohort-ex")
}
