% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncate_window.R
\name{truncate_window}
\alias{truncate_window}
\title{Truncate time window for computation/plot}
\usage{
truncate_window(df, start, end)
}
\arguments{
\item{end}{}
}
\description{
Truncate time window for computation/plot
}
