% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_avg_glucose.R
\name{compute_avg_glucose}
\alias{compute_avg_glucose}
\title{Compute average glucose in given time periods}
\usage{
compute_avg_glucose(df_dex)
}
\arguments{
\item{df_dex}{Data frame outputted by read_dexcom}

\item{start}{Date-time. Computation window start date. Value of "default" means
function ignores this parameter at computes up to the start of the data.}

\item{end}{Date-time. Computation window end date. Value of "default" means
function ignores this parameter at computes up to the start of the data.}
}
\description{
Compute average glucose in given time periods
}
