% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_avg_glucose.R
\name{compute_avg_glucose}
\alias{compute_avg_glucose}
\title{Compute average glucose in given time periods}
\usage{
compute_avg_glucose(df, inter, human_run = T)
}
\arguments{
\item{df}{Data frame, read_dexcom output}

\item{inter}{Time interval for computation; e.g. every 14 days, every month, etc.}

\item{human_run}{Boolean. For internal use, default set to TRUE.}
}
\description{
Compute average glucose in given time periods
}
