% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_gmi.R
\name{compute_gmi}
\alias{compute_gmi}
\title{Compute GMI}
\usage{
compute_gmi(df_dex, start = "default", end = "default", inter, include_bv = T)
}
\arguments{
\item{df_dex}{Data frame outputted by read_dexcom}

\item{start}{Date-time. Computation window start date. Value of "default" means
function ignores this parameter at computes up to the start of the data.}

\item{end}{Date-time. Computation window end date. Value of "default" means
function ignores this parameter at computes up to the start of the data.}

\item{inter}{Time interval for computation; e.g. every 14 days, every month, etc.}

\item{include_bv}{If TRUE, "High" and "Low" flags are converted to value 400 and 40
respectively and used in computations. If FALSE, this conversion does not happen;
computations ignore the "High" and "Low" flags}
}
\description{
Computes the glucose management indicator for a given time frame.
GMI is computed following the results of:
https://diabetesjournals.org/care/article/42/4/e60/36083/Glucose-Management-Indicator-GMI-Insights-and
}
