% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_gmi.R
\name{compute_gmi}
\alias{compute_gmi}
\title{Compute GMI}
\usage{
compute_gmi(x, y = 0, z = 0, exclude_bv = FALSE, allow_norm = TRUE)
}
\arguments{
\item{x}{A data frame (in the form of read_dexcom output)}

\item{y}{Time period length in weeks}

\item{z}{Time period length in days}

\item{exclude_bv}{Boolean. Compute GMI with high/low values as 400/___ respectively?}

\item{allow_norm}{Boolean. Compute GMI with time periods 'rounded' to start of day}
}
\description{
Computes the glucose management indicator for a given time frame.
GMI is computed following the results of:
https://diabetesjournals.org/care/article/42/4/e60/36083/Glucose-Management-Indicator-GMI-Insights-and
}
