% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_wear_time.R
\name{compute_wear_time}
\alias{compute_wear_time}
\title{Compute wear time for individual patient}
\usage{
compute_wear_time(df_dex, start = "default", end = "default", fumonths)
}
\arguments{
\item{df_dex}{Data frame outputted by read_dexcom}

\item{start}{Date of CGM initiation.}

\item{end}{End of study}

\item{fumonths}{Study window of interest (as months since T1D onset)}
}
\value{
Individual wear time
}
\description{
Compute wear time for individual patient
}
