% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_wear_time.R
\name{compute_wear_time}
\alias{compute_wear_time}
\title{Compute wear time for individual patient}
\usage{
compute_wear_time(df_dex, start = "default", end = "default", fumonths)
}
\arguments{
\item{df_dex}{Data frame outputted by read_dexcom}

\item{start}{Date of CGM initiation.}

\item{end}{End of study}

\item{fumonths}{Study window of interest (as months since T1D onset)}
}
\value{
Individual wear time
}
\description{
Compute wear time for individual patient
}
\examples{
bg_date_time <- as.POSIXct(c("2020-01-01 01:01:01", "2020-03-01 01:01:01",
"2020-06-01 01:01:01", "2020-07-01 01:01:01", "2020-09-09 01:01:01", "2020-12-09 01:01:01",
"2021-01-01 01:01:01"), tz = "UTC")
bg_value_num <- c(67, 100, 150, 123, 124, 80, 300)
record_id <- c(010101, 010101, 010101, 010101, 010101, 010101, 010101)
bg_value_flag <- c(NA, NA, NA, NA, NA, NA, NA)


df <- dplyr::tibble(bg_date_time = bg_date_time, bg_value_num = bg_value_num, record_id = record_id,
bg_value_flag = bg_value_flag)
compute_wear_time(df)

}
