% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_agp.R
\name{plot_agp}
\alias{plot_agp}
\title{Canonical bar plot for AGP Thermometers}
\usage{
plot_agp(df, inter)
}
\arguments{
\item{df}{read_dexcom output}

\item{inter}{time interval}
}
\description{
Canonical bar plot for AGP Thermometers
}
