% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_inter.R
\name{set_inter}
\alias{set_inter}
\title{Creates 'inter' column for read_dexcom ouptut data frame}
\usage{
set_inter(x, interval)
}
\arguments{
\item{x}{Input data frame}

\item{interval}{Integer. Time interval, in days.}
}
\value{
Original (input) data frame with new column 'inter'. 'inter' column contains integer
indicating which time interval an observation is from (e.g. 1 indicates from first time interval, 2
indicates from the second, etc.)
}
\description{
Creates 'inter' column for read_dexcom ouptut data frame
}
