% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_inter.R
\name{set_inter}
\alias{set_inter}
\title{Creates 'inter' column for read_dexcom ouptut data frame}
\usage{
set_inter(df_dex, interval)
}
\arguments{
\item{df_dex}{Data frame outputted by read_dexcom}

\item{interval}{Integer. Time interval, in days.}
}
\value{
Input data frame with new column 'inter' which is an integer
indicating which time interval an observation is from
}
\description{
Creates 'inter' column for read_dexcom ouptut data frame
}
